Update OpenFaaS to use helm3 via arkade
arkade always has the most recent and up to date way to install
openfaas. It is the official CLI installer, and can pull in helm3 into
its own private folder $HOME/.arkade/bin/helm for the
installation.

This PR should reduce maintenance overheads - whenever the
install changes, we simply change arkade, and users benefit from
the changes. The password name is the same, the value passed in
from the Civo code works the same too.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>